Implement YotiDocScanSdkStopper
Implemented a way to force close the Yoti SDK from the outside, by introducing a
component that registers for ActivityLifecycleCallbacks and keeps track of the sdk's
activities and is able to call .finish() on them, when requested.

Meant to be used as a last resort only, since such functionality should ideally
be provided by the native SDK instead.

Related to YD-5354