Remove forced androidx core lib version and update compile/target SDK
The androidx.core:core forced lib version was used in the past to fix a build error
which no longer seems to occur.

The downside was it was very outdated, with newer versions of the lib adding
ProtectionLayout, which now caused a ClassNotFoundException.

The solution was to remove the version forcing, so that whatever YDS declares
is used for the build. This also needed a compile and target sdk version update
to 35.

Related to DEP-621